requirements.txt
django==2.2.4
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
psycopg2==2.8.3
Pillow==6.1.0
gunicorn==20.0.4
djangorestframework-simplejwt==4.4.0
